requirements.txt
Django==4.2.7
tensorflow>=2.16.0
Pillow>=10.0.1
numpy>=1.24.3
opencv-python>=4.8.1.78